==> Data/3.  Comtrade — Data wrangling.Rmd <==
# Trade flows dataset

```{r}
pacman::p_load(tidyverse, here)
```

### Loading exports and imports

```{r}
# Exports
setwd(here("Data", "Exports raw"))

exports_raw <- 
  dir(pattern = "*.rds") |> 
  map(read_rds) |> 
  reduce(bind_rows)

# Imports
setwd(here("Data", "Imports raw"))

imports_raw <-
  dir(pattern = "*.rds") |>
  map(read_rds) |>
  reduce(bind_rows)

setwd(here())
```

### Binding and cleaning

```{r}
trade <- 
  bind_rows(exports_raw, imports_raw) |> 
  select_if(~ !any(is.na(.))) |> 
  select(-c(classification, period, period_desc, 
            is_leaf_code, trade_flow_code, flag, 
            aggregate_level, reporter_code, partner_code)) |> 
  group_by(trade_flow) |> 
  arrange(year, trade_flow, partner, commodity) |> 
  mutate(partner = case_when(
    partner == "Dominican Rep." ~ "Dominican Republic",
    partner == "Bolivia (Plurinational State of)" ~ "Bolivia",
    TRUE ~ partner
  )) |> 
  ungroup()
```

### Writing tibble

```{r}
write_csv(trade, file = paste0(here("Data", "Main data frames"), "/trade_flows.csv"))
```
