==> Data/4.  World Bank — Data extraction.Rmd <==
# Macro variables

```{r}
pacman::p_load(tidyverse, here, WDI)
```

### Names query

```{r}
# as_tibble(WDIsearch("exchange")) %>% print(n = Inf)
```

## (a) Extraction function

### Search parameters

```{r}
search <- function(country) {
  data <- WDI(
    indicator = c(
      gdp = "NY.GDP.MKTP.CD", # (current US$)
      gdp_pcap = "NY.GDP.PCAP.CD", # (current US$)
      # expend = "NE.DAB.TOTL.CD", # (current US$)
      population = "SP.POP.TOTL",
      exchange = "PA.NUS.FCRF",
      deflator = "NY.GDP.DEFL.ZS"
    ),
    country = country,
    start = 1995,
    end = 2020,
    extra = TRUE
  ) %>% 
    as_tibble()
}
```

### Export

```{r}
write <-
  function(data) {
    name <- deparse(substitute(data))
    write_csv(data, file = paste0(here("Data", "World Bank raw"), "/", name, ".csv"))
  }
```

## (b) Data extraction

### Countries in the main dataset

```{r}
partners <- 
  read_csv(here("Data", "Main data frames", "partners_names.csv")) %>% 
  pull(partner_iso)
```

### Extraction

```{r}
partners_wdi <- tibble()

for (partner in partners) {
  retrieve <- search(partner)
  partners_wdi <- bind_rows(partners_wdi, retrieve)
}

write(partners_wdi)
```

```{r}
guatemala_wdi <- search("GT")
write(guatemala_wdi)
```
