==> Data/4. Partners — Definition.Rmd <==
# Partners dataset

```{r}
pacman::p_load(tidyverse, here)
```

### Country codes

```{r}
codes <-
  read_csv("https://git.io/v564u") |> 
  select(partner = name,
         partner_iso = `alpha-3`,
         subregions = `sub-region`,
         int_region = `intermediate-region`)|> 
  mutate(partner = case_when(
    partner == "Bolivia (Plurinational State of)" ~ "Bolivia",
    TRUE ~ partner 
  ))
```

### Defining trade partners

United Nations geoscheme: <https://en.wikipedia.org/wiki/United_Nations_geoscheme#Americas>

```{r}
names <- 
  tribble(~partner,
        # North America
        "United States of America", "Canada", "Mexico",
        
        # Central America and Mexico
        "Belize", "Costa Rica", "Panama", 
        "Honduras", "Nicaragua", "El Salvador", 
        
        # South America (-Venezuela)
        "Argentina", "Bolivia",  "Brazil", "Chile",
        "Colombia", "Ecuador", "Guyana", "Paraguay", 
        "Peru", "Uruguay", "Suriname",
        
        # Caribbean independent countries
        # (-Antigua and Barbuda, Cuba, Saint Kitts and Nevis)
        "Bahamas", "Barbados", "Dominica", "Dominican Republic",
        "Grenada", "Haiti", "Jamaica", "Saint Lucia", 
        "Saint Vincent and the Grenadines", "Trinidad and Tobago"
        )
```

```{r}
countries <- 
  codes |> 
  inner_join(names, by = "partner") |> 
  arrange(partner, int_region)
```

### Writing tibble

```{r}
countries |> 
  write_csv(file = paste0(here("Data", "Main data frames"), "/partners_names.csv"))
```
