==> Data/5.  CEPII — Data extraction.Rmd <==
# Transaction costs dataset

```{r}
pacman::p_load(tidyverse, here)
```

## CEPII database

Documentation: <http://www.cepii.fr/DATA_DOWNLOAD/gravity/doc/Gravity_documentation.pdf>

```{r}
gravity_raw <- read_csv(here("Data", "CEPII raw", "Gravity_V202102.csv"))
```

## Transaction cost variables

$$
t_{ij}= D_{ij} \cdot \exp(cont_{ij} + lang_{ij} + ccol_{ij} + crel_{ij} + rta_{ij})
$$

Where $D_{ij}$ is distance measured by the great circle formula, $contig_{ij}$ denotes whether the two countries have a common border, $lang_{ij}$ wheter both countries share the same language, $ccol_{ij}$ both nations have a common colonizer, $crel_{ij}$ is a similarity index between religious practices and $rta_{ij}$ whether the twocountries are members of a regional trade agreement.

```{r}
costs_raw <- 
  gravity_raw |> 
  filter(between(year, 1995, 2019), iso3_o == "GTM") |> 
  select(year, 
         reporter_iso = iso3_o, 
         partner_iso = iso3_d, 
         # Distance
         dist, distw, distcap, distwces,
         # Geographic and cultural variables
         cont = contig,
         lang = comlang_off, 
         ccol = sibling_ever,
         crel = comrelig,
         rta)
```

### Filtering partners

```{r}
partners <- 
  read_csv(here("Data", "Main data frames", "partners_names.csv")) |> 
  select(partner_iso)

costs <- 
  costs_raw |> 
  inner_join(partners)
```

### Adding 2020

```{r}
transaction_costs <- 
  costs |> 
  group_by(partner_iso) |>
  uncount((row_number() == max(row_number())) + 1) |> 
  ungroup() |> 
  mutate(year_2 = rep(1995:2020, times = 30)) |> 
  select(year = year_2, reporter_iso:rta)
```

### Writing tibble

```{r}
transaction_costs |> 
  write_csv(here("Data", "Main data frames", "transaction_costs.csv"))
```
