==> Data/7. World Bank — Data wrangling.Rmd <==
# Macro variables dataset

```{r}
pacman::p_load(tidyverse, here)
```

### Partners

```{r}
countries <- 
  read_csv(here("Data", "Main data frames", "partners_names.csv")) |> 
  pull(partner_iso)
```

```{r}
partners <- 
  read_csv(here("Data", "World Bank raw", "partners_wdi.csv")) |> 
  select(year:iso3c, capital:latitude) |> 
  rename("partner_iso" = iso3c) |> 
  filter(partner_iso %in% countries) |> 
  select(!c(status, lastupdated, capital, longitude, latitude)) |> 
  relocate(year, partner_iso, everything())
```

Adding exchange rate for HND based on: <https://fred.stlouisfed.org/series/FXRATEHNA618NUPN>

```{r}
hnd <- tibble(year = 1995:1999,
              partner_iso = "HND",
              exchange = c(9.470986667, 11.7053025,
                           13.00347417, 13.385015,
                           14.21316833))

partners <- 
  partners |> 
  rows_update(hnd, by = c("year", "partner_iso"))
```

Adding implicit price deflator for CAN based on: <https://fred.stlouisfed.org/series/CANGDPDEFQISMEI>

```{r}
can <- tibble(year = 1995:1996,
              partner_iso = "CAN",
              deflator = c(68.05195, 69.23677))

partners <- 
  partners |> 
  rows_update(can, by = c("year", "partner_iso"))
```

### Guatemala

```{r}
guatemala <- 
  read_csv(here("Data", "World Bank raw", "guatemala_wdi.csv")) |> 
  select(year:iso3c, capital:latitude) |> 
  rename("reporter_iso" = iso3c) |> 
  select(!c(status, lastupdated, capital, longitude, latitude)) |> 
  relocate(year, reporter_iso, everything())
```

### Writing tibbles

```{r}
write_csv(partners, file = paste0(here("Data", "Main data frames"), "/partners_macro.csv"))
write_csv(guatemala, file = paste0(here("Data", "Main data frames"), "/gtm_macro.csv"))
```
