==> Data/8.  Gravity — Data wrangling.Rmd <==
# Gravity dataset

```{r}
pacman::p_load(tidyverse, here)
```

```{r}
gtm_macro <- read_csv(here("Data", "Main data frames", "gtm_macro.csv"))
partners_macro <- read_csv(here("Data", "Main data frames", "partners_macro.csv"))
trade_flows <- read_csv(here("Data", "Main data frames", "trade_flows.csv"))
transaction_costs <- read_csv(here("Data", "Main data frames", "transaction_costs.csv"))
```

## Gravity variables

$$
TGDP_{ij,t} = ln(GDP_{i,t} + GDP_{j,t}) \\
RFE_{ij, t} = \Bigg| ln\Bigg(\frac{GDP_{i, t}}{N_{i, t}}\Bigg) - ln\Bigg(\frac{GDP_{j, t}}{N_{j, t}}\Bigg) \Bigg| \\
SIM_{ij, t} = 1 - \Big(\frac{GDP_{i, t}}{GDP_{i, t}+GDP_{j, t}}\Big)^2 - \Big(\frac{GDP_{j, t}}{GDP_{i, t}+GDP_{j, t}}\Big)^2
$$

```{r}
macro <- 
  gtm_macro |> 
  left_join(partners_macro, by = "year", suffix = c("_i", "_j")) |> 
  arrange(partner_iso, year) |> 
  group_by(partner_iso) |> 
  mutate(
    tgdp = log(gdp_i + gdp_j),
    rfe = abs(log(gdp_i/population_i) - log(gdp_j/population_j)),
    sim = 1 - (gdp_i/(gdp_i + gdp_j))^2 - (gdp_j/(gdp_i + gdp_j))^2
  ) |> 
  ungroup()
```

## Gravity dataset

### Exports

```{r}
total_exports <- 
  trade_flows |> 
  filter(trade_flow == "Export") |> 
  group_by(year, partner_iso) |> 
  summarise(trade_value = sum(trade_value_usd)) |> 
  ungroup() |> 
  arrange(partner_iso)
```

```{r}
# Countries and years for which there are zero trade flows
transaction_costs |> anti_join(total_exports) |> print(n = Inf)
```

```{r}
exports <- 
  transaction_costs |> 
  left_join(total_exports) |> 
  select(year:partner_iso, trade_value, everything()) |> 
  mutate(across(where(is.numeric), ~ replace(., is.na(.), 0))) |> 
  left_join(macro, by = c("year", "reporter_iso", "partner_iso")) |> 
  rename("iso_i" = reporter_iso, "iso_j" = partner_iso, "exports" = trade_value)
```

### Imports

```{r}
total_imports <- 
  trade_flows |> 
  filter(trade_flow == "Import") |> 
  group_by(year, partner_iso) |> 
  summarise(trade_value = sum(trade_value_usd)) |> 
  ungroup() |> 
  arrange(partner_iso)
```

```{r}
# Zeros in the total_import data set
transaction_costs |> anti_join(total_imports) 
```

```{r}
imports <- 
  transaction_costs |> 
  left_join(total_imports) |> 
  select(year:partner_iso, trade_value, everything()) |> 
  mutate(across(where(is.numeric), ~ replace(., is.na(.), 0))) |> 
  left_join(macro, by = c("year", "reporter_iso", "partner_iso")) |> 
  rename("iso_i" = reporter_iso, "iso_j" = partner_iso, "imports" = trade_value)
```

```{r}
# Correcting 0's due to errors in the Comtrade API
can <- tibble(year = c(1999, 2000, 2002, 2005, 2008, 2013, 2018),
              iso_j = "CAN",
              imports = c(118894015, 120584218, 69796525, 190471222,
                          174517492, 86239715, 102364873))

imports <- 
  imports |> 
  rows_update(can, by = c("year", "iso_j"))
```

### Gravity

```{r}
gravity <- 
  exports |> 
  select(exports) |>
  bind_cols(imports) |> 
  select(year:iso_j, exports, imports, 
         tgdp, rfe, sim, dist, cont, lang, ccol, crel, rta) |> 
  drop_na() |> 
  write_csv("gravity.csv")
```
